clean up sgl kernel
